add minimum column width
