+++ src/ReactGrid.tsx
@@ -26,6 +26,7 @@ const ReactGrid: FC<ReactGridProps> = ({
   styledRanges,
   initialSelectedRange,
   initialFocusLocation,
+  minColumnWidth,
   onAreaSelected,
   onFillHandle,
   onCellFocused,
@@ -41,6 +42,7 @@ const ReactGrid: FC<ReactGridProps> = ({
     behaviors,
     userStyles,
     styledRanges,
+    minColumnWidth,
     onFillHandle,
     onAreaSelected,
     onCellFocused,
