+++ src/stories/utils/onPasteHandler.ts
@@ -1,0 +1,20 @@
+import { NumericalRange } from "../../types/CellMatrix";
+
+export const onPasteHandler = <T>(
+  selectedArea: NumericalRange,
+  pastedData: string,
+  setData: React.Dispatch<React.SetStateAction<T[][]>>
+) => {
+  // parse the pasted data
+  const parsedData = JSON.parse(pastedData);
+
+  setData((prev) => {
+    const next = [...prev];
+    for (let i = 0; i < parsedData.length; i++) {
+      for (let j = 0; j < parsedData[i].length; j++) {
+        next[selectedArea.startRowIdx + i][selectedArea.startColIdx + j] = parsedData[i][j];
+      }
+    }
+    return next;
+  });
+};
